requirements.txt
# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.1

# Redis
redis>=5.0.1

# Telegram Bot API
aiogram>=3.4.1
aiohttp>=3.9.1

# Settings & Validation
pydantic>=2.5.3
pydantic-settings>=2.1.0

# Scheduler
apscheduler>=3.10.4

# Logging
structlog>=24.1.0
python-json-logger>=2.0.7

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
httpx>=0.26.0

# Development
black>=23.12.1
ruff>=0.1.11
